+++ src/C3.WinForms/Features/CassetteModels/frmModelNew.vb
@@ -27,7 +27,7 @@ Public Class frmModelNew
         CompleteCatalogueMutation(Me)
 
         Dim message As String = "Added cassette model " & result.Model.Identifier & " successfully."
-        If My.Settings.showMessages Then
+        If preferences.ShowMessages Then
             MsgBox(message, MsgBoxStyle.Information, "Cassette Model Added")
         End If
         consoleAdd(message)